refactor(files): moves documents into separate folder
