+++ schemas/documents/blogSeries.tsx
