+++ schemas/index.ts
@@ -1,10 +1,10 @@
 import {about} from './pages/about'
-import {blogPost} from './blogPost'
-import {testimonial} from './testimonial'
+import {blogPost} from './documents/blogPost'
+import {testimonial} from './documents/testimonial'
 import {contentComponent} from './components/content'
-import {blogSeries} from './blogSeries'
+import {blogSeries} from './documents/blogSeries'
 import {resume} from './pages/resume'
-import {project} from './project'
+import {project} from './documents/project'
 import {home} from './pages/home'
 import {buttonComponent} from './components/button'
